DRILL-4606: HYGIENE
+ Merge DrillAutoCloseables and AuthCloseables
+ Remove unused imports
+ Expand * imports
